refactor: cardInfoStateInterface의 위치 변경
