--- src/context/CardInfoContext.tsx
@@ -1,8 +1,0 @@
-export interface CardInfoStateTypeInterface {
-  cardNumber: string[];
-  holderName: string;
-  expireDate: string[];
-  securityCodeLength: number;
-  passwordLength: number[];
-  nickname?: string;
-}
